Enhanced README with build instructions, replaced int with uint8_t/int16_t and added CMakeLists.txt for project configuration

+++ include/utility_functions.h
@@ -11,8 +11,8 @@
 typedef struct
 {
     unsigned char data[TREE_LEVELS][SEED_LENGTH]; // array of seeds
-    int levels[TREE_LEVELS];                      // array of levels, to memorize the seed's height
-    int top;
+    uint8_t levels[TREE_LEVELS];                  // array of levels, to memorize the seed's height
+    int16_t top;
 } SeedStack;
 
 // saves the tree state across different function calls
@@ -20,9 +20,9 @@ typedef struct
 {
     SeedStack stack;
     unsigned char *seed;
-    int *bitmask;
-    int bitmaskSize;
-    unsigned int iterations;
+    uint8_t *bitmask;
+    uint8_t bitmaskSize;
+    uint8_t iterations;
 } TreeData;
 
 // debug, given the seed x returns [2x+1, 2x+2]
@@ -37,5 +37,5 @@ void leftSeed(unsigned char *output, const unsigned char *input);
 // copies the second half of the parent seed
 void rightSeed(unsigned char *output, const unsigned char *input);
 
-void push(SeedStack *s, const unsigned char *seed, const int seedLevel);
+void push(SeedStack *s, const unsigned char *seed, const uint8_t seedLevel);
 unsigned char *pop(SeedStack *s);
